Refactoring: move spring boot logging to spring boot common
