+++ src/main/java/tech/jhipster/lite/generator/server/springboot/common/application/SpringBootCommonApplicationService.java
@@ -2,6 +2,7 @@ package tech.jhipster.lite.generator.server.springboot.common.application;
 
 import org.springframework.stereotype.Service;
 import tech.jhipster.lite.generator.project.domain.Project;
+import tech.jhipster.lite.generator.server.springboot.common.domain.Level;
 import tech.jhipster.lite.generator.server.springboot.common.domain.SpringBootCommonService;
 
 @Service
@@ -27,5 +28,13 @@ public class SpringBootCommonApplicationService {
 
   public void addPropertiesTest(Project project, String key, Object value) {
     springBootCommonService.addPropertiesTest(project, key, value);
+  }
+
+  public void addLogger(Project project, String packageName, Level level) {
+    springBootCommonService.addLogger(project, packageName, level);
+  }
+
+  public void addLoggerTest(Project project, String packageName, Level level) {
+    springBootCommonService.addLoggerTest(project, packageName, level);
   }
 }
