+++ src/main/java/tech/jhipster/lite/generator/server/springboot/database/mysql/domain/MySQLDomainService.java
@@ -13,9 +13,8 @@ import tech.jhipster.lite.generator.buildtool.generic.domain.Dependency;
 import tech.jhipster.lite.generator.project.domain.DefaultConfig;
 import tech.jhipster.lite.generator.project.domain.Project;
 import tech.jhipster.lite.generator.project.domain.ProjectRepository;
+import tech.jhipster.lite.generator.server.springboot.common.domain.Level;
 import tech.jhipster.lite.generator.server.springboot.common.domain.SpringBootCommonService;
-import tech.jhipster.lite.generator.server.springboot.logging.domain.Level;
-import tech.jhipster.lite.generator.server.springboot.logging.domain.SpringBootLoggingService;
 
 public class MySQLDomainService implements MySQLService {
 
@@ -24,18 +23,15 @@ public class MySQLDomainService implements MySQLService {
   private final ProjectRepository projectRepository;
   private final BuildToolService buildToolService;
   private final SpringBootCommonService springBootCommonService;
-  private final SpringBootLoggingService springBootLoggingService;
 
   public MySQLDomainService(
     ProjectRepository projectRepository,
     BuildToolService buildToolService,
-    SpringBootCommonService springBootCommonService,
-    SpringBootLoggingService springBootLoggingService
+    SpringBootCommonService springBootCommonService
   ) {
     this.projectRepository = projectRepository;
     this.buildToolService = buildToolService;
     this.springBootCommonService = springBootCommonService;
-    this.springBootLoggingService = springBootLoggingService;
   }
 
   @Override
@@ -103,8 +99,8 @@ public class MySQLDomainService implements MySQLService {
     addLogger(project, "org.hibernate", Level.WARN);
     addLogger(project, "org.hibernate.ejb.HibernatePersistence", Level.OFF);
 
-    springBootLoggingService.addLoggerTest(project, "com.github.dockerjava", Level.WARN);
-    springBootLoggingService.addLoggerTest(project, "org.testcontainers", Level.WARN);
+    springBootCommonService.addLoggerTest(project, "com.github.dockerjava", Level.WARN);
+    springBootCommonService.addLoggerTest(project, "org.testcontainers", Level.WARN);
   }
 
   @Override
@@ -161,7 +157,7 @@ public class MySQLDomainService implements MySQLService {
   }
 
   public void addLogger(Project project, String packageName, Level level) {
-    springBootLoggingService.addLogger(project, packageName, level);
-    springBootLoggingService.addLoggerTest(project, packageName, level);
+    springBootCommonService.addLogger(project, packageName, level);
+    springBootCommonService.addLoggerTest(project, packageName, level);
   }
 }
