+++ src/main/java/tech/jhipster/lite/generator/server/springboot/database/mysql/infrastructure/config/MySQLBeanConfiguration.java
@@ -7,7 +7,6 @@ import tech.jhipster.lite.generator.project.domain.ProjectRepository;
 import tech.jhipster.lite.generator.server.springboot.common.domain.SpringBootCommonService;
 import tech.jhipster.lite.generator.server.springboot.database.mysql.domain.MySQLDomainService;
 import tech.jhipster.lite.generator.server.springboot.database.mysql.domain.MySQLService;
-import tech.jhipster.lite.generator.server.springboot.logging.domain.SpringBootLoggingService;
 
 @Configuration
 public class MySQLBeanConfiguration {
@@ -15,22 +14,19 @@ public class MySQLBeanConfiguration {
   public final ProjectRepository projectRepository;
   public final BuildToolService buildToolService;
   public final SpringBootCommonService springBootCommonService;
-  public final SpringBootLoggingService springBootLoggingService;
 
   public MySQLBeanConfiguration(
     ProjectRepository projectRepository,
     BuildToolService buildToolService,
-    SpringBootCommonService springBootCommonService,
-    SpringBootLoggingService springBootLoggingService
+    SpringBootCommonService springBootCommonService
   ) {
     this.projectRepository = projectRepository;
     this.buildToolService = buildToolService;
     this.springBootCommonService = springBootCommonService;
-    this.springBootLoggingService = springBootLoggingService;
   }
 
   @Bean
   public MySQLService mysqlService() {
-    return new MySQLDomainService(projectRepository, buildToolService, springBootCommonService, springBootLoggingService);
+    return new MySQLDomainService(projectRepository, buildToolService, springBootCommonService);
   }
 }
