+++ src/main/java/tech/jhipster/lite/generator/server/springboot/database/postgresql/domain/PostgresqlDomainService.java
@@ -13,9 +13,8 @@ import tech.jhipster.lite.generator.buildtool.generic.domain.BuildToolService;
 import tech.jhipster.lite.generator.buildtool.generic.domain.Dependency;
 import tech.jhipster.lite.generator.project.domain.Project;
 import tech.jhipster.lite.generator.project.domain.ProjectRepository;
+import tech.jhipster.lite.generator.server.springboot.common.domain.Level;
 import tech.jhipster.lite.generator.server.springboot.common.domain.SpringBootCommonService;
-import tech.jhipster.lite.generator.server.springboot.logging.domain.Level;
-import tech.jhipster.lite.generator.server.springboot.logging.domain.SpringBootLoggingService;
 
 public class PostgresqlDomainService implements PostgresqlService {
 
@@ -24,18 +23,15 @@ public class PostgresqlDomainService implements PostgresqlService {
   private final ProjectRepository projectRepository;
   private final BuildToolService buildToolService;
   private final SpringBootCommonService springBootCommonService;
-  private final SpringBootLoggingService springBootLoggingService;
 
   public PostgresqlDomainService(
     ProjectRepository projectRepository,
     BuildToolService buildToolService,
-    SpringBootCommonService springBootCommonService,
-    SpringBootLoggingService springBootLoggingService
+    SpringBootCommonService springBootCommonService
   ) {
     this.projectRepository = projectRepository;
     this.buildToolService = buildToolService;
     this.springBootCommonService = springBootCommonService;
-    this.springBootLoggingService = springBootLoggingService;
   }
 
   @Override
@@ -127,8 +123,8 @@ public class PostgresqlDomainService implements PostgresqlService {
     addLogger(project, "org.hibernate.ejb.HibernatePersistence", Level.OFF);
     addLogger(project, "org.postgresql", Level.WARN);
 
-    springBootLoggingService.addLoggerTest(project, "com.github.dockerjava", Level.WARN);
-    springBootLoggingService.addLoggerTest(project, "org.testcontainers", Level.WARN);
+    springBootCommonService.addLoggerTest(project, "com.github.dockerjava", Level.WARN);
+    springBootCommonService.addLoggerTest(project, "org.testcontainers", Level.WARN);
   }
 
   private Map<String, Object> springPropertiesDatasource(String baseName) {
@@ -195,7 +191,7 @@ public class PostgresqlDomainService implements PostgresqlService {
   }
 
   public void addLogger(Project project, String packageName, Level level) {
-    springBootLoggingService.addLogger(project, packageName, level);
-    springBootLoggingService.addLoggerTest(project, packageName, level);
+    springBootCommonService.addLogger(project, packageName, level);
+    springBootCommonService.addLoggerTest(project, packageName, level);
   }
 }
