--- src/main/java/tech/jhipster/lite/generator/server/springboot/logging/domain/SpringBootLoggingService.java
@@ -1,9 +1,0 @@
-package tech.jhipster.lite.generator.server.springboot.logging.domain;
-
-import tech.jhipster.lite.generator.project.domain.Project;
-
-public interface SpringBootLoggingService {
-  void addLogger(Project project, String packageName, Level level);
-
-  void addLoggerTest(Project project, String packageName, Level level);
-}
