--- src/main/java/tech/jhipster/lite/generator/server/springboot/logging/package-info.java
@@ -1,2 +1,0 @@
-@tech.jhipster.lite.SharedKernel
-package tech.jhipster.lite.generator.server.springboot.logging;
