+++ src/main/java/tech/jhipster/lite/generator/server/springboot/mvc/web/infrastructure/config/SpringBootMvcBeanConfiguration.java
@@ -5,7 +5,6 @@ import org.springframework.context.annotation.Configuration;
 import tech.jhipster.lite.generator.buildtool.generic.domain.BuildToolService;
 import tech.jhipster.lite.generator.project.domain.ProjectRepository;
 import tech.jhipster.lite.generator.server.springboot.common.domain.SpringBootCommonService;
-import tech.jhipster.lite.generator.server.springboot.logging.domain.SpringBootLoggingService;
 import tech.jhipster.lite.generator.server.springboot.mvc.web.domain.SpringBootMvcDomainService;
 import tech.jhipster.lite.generator.server.springboot.mvc.web.domain.SpringBootMvcService;
 
@@ -15,22 +14,19 @@ public class SpringBootMvcBeanConfiguration {
   public final ProjectRepository projectRepository;
   public final BuildToolService buildToolService;
   public final SpringBootCommonService springBootCommonService;
-  public final SpringBootLoggingService springBootLoggingService;
 
   public SpringBootMvcBeanConfiguration(
     ProjectRepository projectRepository,
     BuildToolService buildToolService,
-    SpringBootCommonService springBootCommonService,
-    SpringBootLoggingService springBootLoggingService
+    SpringBootCommonService springBootCommonService
   ) {
     this.projectRepository = projectRepository;
     this.buildToolService = buildToolService;
     this.springBootCommonService = springBootCommonService;
-    this.springBootLoggingService = springBootLoggingService;
   }
 
   @Bean
   public SpringBootMvcService springBootMvcService() {
-    return new SpringBootMvcDomainService(projectRepository, buildToolService, springBootCommonService, springBootLoggingService);
+    return new SpringBootMvcDomainService(projectRepository, buildToolService, springBootCommonService);
   }
 }
