+++ src/test/java/tech/jhipster/lite/generator/server/springboot/common/domain/SpringBootCommonDomainServiceTest.java
@@ -1,15 +1,15 @@
 package tech.jhipster.lite.generator.server.springboot.common.domain;
 
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.verify;
 import static tech.jhipster.lite.TestUtils.tmpProject;
+import static tech.jhipster.lite.TestUtils.tmpProjectWithSpringBootLoggingConfiguration;
 import static tech.jhipster.lite.common.domain.FileUtils.getPath;
 import static tech.jhipster.lite.common.domain.FileUtils.getPathOf;
 import static tech.jhipster.lite.generator.project.domain.Constants.MAIN_RESOURCES;
 import static tech.jhipster.lite.generator.project.domain.Constants.TEST_RESOURCES;
-import static tech.jhipster.lite.generator.server.springboot.core.domain.SpringBoot.APPLICATION_FAST_PROPERTIES;
-import static tech.jhipster.lite.generator.server.springboot.core.domain.SpringBoot.APPLICATION_PROPERTIES;
+import static tech.jhipster.lite.generator.server.springboot.core.domain.SpringBoot.*;
+import static tech.jhipster.lite.generator.server.springboot.core.domain.SpringBoot.NEEDLE_LOGBACK_LOGGER;
 
 import java.nio.file.Files;
 import org.junit.jupiter.api.Test;
@@ -72,5 +72,37 @@ class SpringBootCommonDomainServiceTest {
     springBootCommonDomainService.addPropertiesTest(project, "server.port", 8080);
 
     verify(projectRepository).replaceText(any(Project.class), anyString(), anyString(), anyString(), anyString());
+  }
+
+  @Test
+  void shouldAddLogger() {
+    Project project = tmpProjectWithSpringBootLoggingConfiguration();
+
+    springBootCommonDomainService.addLogger(project, "tech.jhipster.lite", Level.ERROR);
+
+    verify(projectRepository)
+      .replaceText(
+        any(Project.class),
+        contains("main"),
+        eq(LOGGING_CONFIGURATION),
+        eq(NEEDLE_LOGBACK_LOGGER),
+        contains(NEEDLE_LOGBACK_LOGGER)
+      );
+  }
+
+  @Test
+  void shouldAddLoggerTest() {
+    Project project = tmpProjectWithSpringBootLoggingConfiguration();
+
+    springBootCommonDomainService.addLoggerTest(project, "tech.jhipster.lite", Level.ERROR);
+
+    verify(projectRepository)
+      .replaceText(
+        any(Project.class),
+        contains("test"),
+        eq(LOGGING_TEST_CONFIGURATION),
+        eq(NEEDLE_LOGBACK_LOGGER),
+        contains(NEEDLE_LOGBACK_LOGGER)
+      );
   }
 }
